backend/requirements.txt
fastapi==0.115.9
uvicorn==0.27.0
sentence-transformers==2.2.2
chromadb==1.0.12
google-generativeai>=0.7.2
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
huggingface-hub==0.24.0
sqlalchemy==2.0.25
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
